requirements.txt
pytest
numpy==1.26.4
scipy==1.12.0
networkx
metis
ortools
tqdm
pyyaml
nodevectors
scikit-learn
k-means-constrained
line_profiler
infomap
leidenalg
python-igraph
